[Fix] Modify way of updating timer timer
It's now based on the difference between current time and last time we
updated the timer. This modification fix a problem due to background
limitation since Android Oreo which stop background task in Doze mode.
Our runnable was paussed so the time wasn't updated. Now, when runnable
automatically restart when device goes out of Doze mode, we do a time
difference and we apply the good new time.